Change reading experience after feedback from gig
This version has downsized the "curating" process considerably,
based on the usability of the program at the gig on April 28th.
It seems simpler to just preview the results in text form, and
then read through them one by one (with the user having the
option to skip articles or exit entirely, during the reading).
This also makes it easier to time the on/off of the program,
and makes the article limit number obsolete.
So, the curating method was a good learning experience but
not necessary in the long term for the program. I would not
have thought that pausing between articles would be good, but
it seems, on reflection, to be a good thing.